refactor(network): move conversion unctions to module
